// graphql-api/lib/interfaces/merge-report.interface.ts
export interface ISummary {
  TradelineSummary: ITradelineSummary;
  InquirySummary: IInquirySummary;
  PublicRecordSummary: IPublicRecordSummary;
  PortfolioCreditSummary: IPortfolioCreditSummary[];
  AccountHistorySummary: IAccountHistorySummary[];
}
export interface ITradelineSummary {
  Experian: ITradelineSummaryInfo;
  Equifax: ITradelineSummaryInfo;
  TransUnion: ITradelineSummaryInfo;
  Merge: ITradelineSummaryInfo;
}
export interface ITradelineSummaryInfo {
  RecentDeliquencyMOP: ICodeRef;
  TotalHistoricalNegatives: number | string | null;
  OpenCollectionAccounts: number | string | null;
  TotalCollectionAccounts: number | string | null;
  HistoricalNegativeAccounts: number | string | null;
  TotalInstallmentAccounts: number | string | null;
  OpenInstallmentAccounts: number | string | null;
  TotalOtherAccounts: number | string | null;
  OpenOtherAccounts: number | string | null;
  OpenMortgageAccounts: number | string | null;
  RecentDeliquencyDate: string | null;
  TotalMortgageAccounts: number | string | null;
  DelinquentAccounts: number | string | null;
  DerogatoryAccounts: number | string | null;
  CloseAccounts: number | string | null;
  TotalAccounts: number | string | null;
  OpenAccounts: number | string | null;
  TotalRevolvingAccounts: number | string | null;
  OpenRevolvingAccounts: number | string | null;
  CreditSummaryPeriod: string | null;
  TotalBalances: number | string | null;
  TotalMonthlyPayments: number | string | null;
  AgeofCredit: number | string | null;
}
export interface IInquirySummary {
  Experian: IInquirySummaryInfo;
  Equifax: IInquirySummaryInfo;
  TransUnion: IInquirySummaryInfo;
  Merge: IInquirySummaryInfo;
}
export interface IInquirySummaryInfo {
  NumberInLast2Years: number | string | null;
}
export interface IPublicRecordSummary {
  Experian: IPublicRecordSummaryInfo;
  Equifax: IPublicRecordSummaryInfo;
  Transunion: IPublicRecordSummaryInfo;
  Merge: IPublicRecordSummaryInfo;
}
export interface IPublicRecordSummaryInfo {
  NumberOfRecords: number | string | null;
}
export interface IPortfolioCreditSummary {
  Transunion: IPortfolioCreditSummaryInfo;
}
export interface IPortfolioCreditSummaryInfo {
  SummaryType: ICodeRef;
  CurrentPaymentDueAmount: number | string | null;
  PriorPaymentDueAmount: number | string | null;
  CurrentActualPaymentAmount: number | string | null;
  PastDueAmount: number | string | null;
  CreditLimitAmount: number | string | null;
  BalanceAmount: number | string | null;
}
export interface IAccountHistorySummary {
  Transunion: IAccountHistorySummaryInfo;
}
export interface IAccountHistorySummaryInfo {
  SummaryType: ICodeRef;
  TotalPaymentRatio: number | string | null;
  ActualPaymentAmount: number | string | null;
  PaymentDueAmount: number | string | null;
  TransactorRevolverIndicator: string | null;
  EndingBalanceAmount: number | string | null;
  AggregateExcessPaymentAmount: number | string | null;
  ActiveAccounts: number | string | null;
  OpenAccounts: number | string | null;
  TimePeriod: string | null;
  EstimatedSpendAmount: number | string | null;
  PriorMonthBalance: number | string | null;
  CreditLimitAmount: number | string | null;
}